+++ production/core/entity/common/src/main/java/module-info.java
@@ -1,0 +1,3 @@
+module template.core.entity.common {
+    exports template.core.entity.common;
+}