+++ production/core/entity/common/src/main/java/template/core/entity/common/Issue.java
@@ -1,0 +1,7 @@
+package template.core.entity.common;
+
+/**
+ * Issue represents an individual important topic or problem that needs to be addressed.
+ */
+public interface Issue {
+}